requirements.txt
django
djangorestframework
pep8
python-dotenv
psycopg2
psycopg2-binary
